rent commit 2.0

+++ app/src/main/java/com/example/rentapp/models/car_models/FastSelectPlans.kt
@@ -4,6 +4,6 @@ import java.io.Serializable
 
 data class FastSelectPlans(
     val price: Int,
-    val days: Int,
+    val days: Long,
     val name: String
 ): Serializable
